requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coverage==7.5.4
Django==4.2.13
django-allauth==0.63.1
django-cleanup==8.1.0
django-debug-toolbar==4.2.0
idna==3.7
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
